Refactor BMW base entity
