+++ homeassistant/components/bmw_connected_drive/binary_sensor.py
@@ -22,9 +22,9 @@ from homeassistant.core import HomeAssistant, callback
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 from homeassistant.util.unit_system import UnitSystem
 
-from . import BMWBaseEntity
 from .const import DOMAIN, UNIT_MAP
 from .coordinator import BMWDataUpdateCoordinator
+from .entity import BMWBaseEntity
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -241,9 +241,8 @@ class BMWBinarySensor(BMWBaseEntity, BinarySensorEntity):
         self._attr_is_on = self.entity_description.value_fn(self.vehicle)
 
         if self.entity_description.attr_fn:
-            self._attr_extra_state_attributes = dict(
-                self._attrs,
-                **self.entity_description.attr_fn(self.vehicle, self._unit_system),
+            self._attr_extra_state_attributes = self.entity_description.attr_fn(
+                self.vehicle, self._unit_system
             )
 
         super()._handle_coordinator_update()
