+++ homeassistant/components/bmw_connected_drive/lock.py
@@ -15,9 +15,9 @@ from homeassistant.core import HomeAssistant, callback
 from homeassistant.exceptions import HomeAssistantError
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-from . import BMWBaseEntity
 from .const import DOMAIN
 from .coordinator import BMWDataUpdateCoordinator
+from .entity import BMWBaseEntity
 
 DOOR_LOCK_STATE = "door_lock_state"
 _LOGGER = logging.getLogger(__name__)
@@ -97,8 +97,6 @@ class BMWLock(BMWBaseEntity, LockEntity):
     def _handle_coordinator_update(self) -> None:
         """Handle updated data from the coordinator."""
         _LOGGER.debug("Updating lock data of %s", self.vehicle.name)
-        # Set default attributes
-        self._attr_extra_state_attributes = self._attrs
 
         # Only update the HA state machine if the vehicle reliably reports its lock state
         if self.door_lock_state_available:
@@ -106,8 +104,8 @@ class BMWLock(BMWBaseEntity, LockEntity):
                 LockState.LOCKED,
                 LockState.SECURED,
             }
-            self._attr_extra_state_attributes["door_lock_state"] = (
-                self.vehicle.doors_and_windows.door_lock_state.value
-            )
+            self._attr_extra_state_attributes = {
+                DOOR_LOCK_STATE: self.vehicle.doors_and_windows.door_lock_state.value
+            }
 
         super()._handle_coordinator_update()
